requirements.txt
streamlit
google-cloud-aiplatform
google-cloud-secret-manager